progress

+++ src/app/auth.service.ts
@@ -8,7 +8,7 @@ import {
 } from "@angular/router";
 import { HttpClient } from "@angular/common/http";
 import { environment } from "../environments/environment";
-
+import swal from 'sweetalert2'
 export interface userToken {
   // Request
   grant_type?: string;
@@ -29,8 +29,28 @@ export class AuthService {
   profileValid = false;
 
   logout(): void {
-    localStorage.clear();
-    window.location.reload();
+
+    swal({
+      title: "خارج می‌شوید؟",
+      text: "تمام اطلاعات ذخیره شده به صورت آفلاین، حذف خواهند شد.",
+      type: "warning",
+      showCancelButton: true,
+
+      preConfirm: () => {
+        return new Promise((resolve, reject) => {
+          swal.showLoading();
+          swal.getConfirmButton().innerText = "در حال خروج"
+
+          localStorage.clear();
+          window.location.reload();
+
+          setTimeout(() => {
+            resolve();
+          }, 1000);
+        });
+      }
+    });
+
   }
 
   loggedIn = false;
